requirements.txt
### CMake ###
CMake

### Flask Dependencies ###
Flask==2.0.3
flask_wtf
werkzeug==2.0.3
wtforms==2.3.3
flask-login
imageio
Flask_SocketIO==4.3.1
Flask-Session==0.3.2
python-engineio==3.13.2
python-socketio==4.6.0

### opencv ###
opencv-python
opencv-contrib-python
onnx
onnx_tf
onnxruntime
tensorflow==2.8.0
tensorflow_addons==0.16.1
tensorflow-estimator==2.12.0
tensorflow_io_gcs_filesystem==0.25.0
tensorflow_probability==0.16.0
imutils
open-faces
keras==2.8.0
face_recognition

### gestureRecognition ###
albumentations
omegaconf
torchvision
pyyaml
pillow

### Math/Plotting ###
numpy
pandas
matplotlib
Pillow
scipy

### eduVid ###
nose
nolearn
mediapipe
transformers
torch
faster-whisper
moviepy


### Data ###
scikit-learn

### Database ###
uuid
pymongo==4.1.1
dnspython

### aditional ###
protobuf==3.20.*
